Add likes and friendships modules with enhanced functionality

+++ films/src/main/java/ru/yandex/practicum/filmorate/films/domain/port/UpdateFilmCommand.java
@@ -2,7 +2,7 @@ package ru.yandex.practicum.filmorate.films.domain.port;
 
 import java.time.LocalDate;
 
-public record UpdateFilmCommand(Integer id,
+public record UpdateFilmCommand(Long id,
                                 String name,
                                 String description,
                                 LocalDate releaseDate,
